test: filter_restrict
