--- cypress/e2e_cucumber/filter_restrict/index.js
@@ -1,1 +1,0 @@
-'../common/index.js'
